requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
blinker==1.6.3
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
cryptography==41.0.3
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
Flask==3.0.3
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.1
frozenlist==1.4.1
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.17.0
MarkupSafe==2.1.3
msgpack==1.0.7
multidict==6.0.5
numpy==1.26.1
oauthlib==3.2.2
openai==0.28.0
psutil==5.9.6
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
roundrobin==0.0.4
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
